modified:   go.mod 	modified:   go.sum 	modified:   main.go 	new file:   mods/ip2info/ip2info.go 	modified:   mods/webserver/webserver.go

+++ mods/webserver/webserver.go
@@ -13,7 +13,7 @@ type Type_ServerStatus struct {
 	Requests        int64
 	Errors          int64
 	StartTime       int
-	ActiveClients   int64
+	ActiveClients   int
 	IPs             map[string]int64
 }
